Reintroduce tests for <Directives/>
